third_party/blink/web_tests/external/wpt/editing/run/delete_6001-7000-expected.txt

This is a testharness.js-based test.
Found 30 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["delete",""]] "<ol><li>fo[o</ol><ol><li>b]ar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foar</li></ol>" but got "<ol><li>fo</li><li>ar</li></ol>"
[FAIL] [["stylewithcss","true"],["delete",""]] "<span style=display:block>fo[o</span><span style=display:block>b]ar</span>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["delete",""]] "<span style=display:none>fo[o</span><span style=display:none>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"display:none\\">fo</span><span style=\\"display:none\\">ar</span>" but got ""
[FAIL] [["stylewithcss","true"],["delete",""]] "<quasit style=display:block>fo[o</quasit><quasit style=display:block>b]ar</quasit>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<quasit style=\\"display:block\\">foar</quasit>" but got "<quasit style=\\"display:block\\">fo<span style=\\"text-align:inherit\\">ar</span></quasit>"
[FAIL] [["stylewithcss","false"],["delete",""]] "<quasit style=display:block>fo[o</quasit><quasit style=display:block>b]ar</quasit>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<quasit style=\\"display:block\\">foar</quasit>" but got "<quasit style=\\"display:block\\">fo<span style=\\"text-align:inherit\\">ar</span></quasit>"
[FAIL] [["delete",""]] "<ol><li>foo</ol>{}<br><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol>" but got "<ol><li>foo</li></ol><ol><li>bar</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<ol><li>foo</ol><p>{}<br></p><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol>" but got "<ol><li>foo</li></ol><ol><li>bar</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<ol><li>foo</ol><p>{}<br></p><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol>" but got "<ol><li>foo</li></ol><ol><li>bar</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<ol><li><p>foo</ol><p>{}<br></p><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p></li><li>bar</li></ol>" but got "<ol><li><p>foo</p></li></ol><ol><li>bar</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<ol><li><p>foo</ol><p>{}<br></p><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p></li><li>bar</li></ol>" but got "<ol><li><p>foo</p></li></ol><ol><li>bar</li></ol>"
[FAIL] [["delete",""]] "<ol id=a><li>foo</ol>{}<br><ol><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol id=\\"a\\"><li>foo</li><li>bar</li></ol>" but got "<ol id=\\"a\\"><li>foo</li></ol><ol><li>bar</li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo</ol>{}<br><ol id=b><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>bar</li></ol>" but got "<ol><li>foo</li></ol><ol id=\\"b\\"><li>bar</li></ol>"
[FAIL] [["delete",""]] "<ol id=a><li>foo</ol>{}<br><ol id=b><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol id=\\"a\\"><li>foo</li><li>bar</li></ol>" but got "<ol id=\\"a\\"><li>foo</li></ol><ol id=\\"b\\"><li>bar</li></ol>"
[FAIL] [["delete",""]] "<ol class=a><li>foo</ol>{}<br><ol class=b><li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol class=\\"a\\"><li>foo</li><li>bar</li></ol>" but got "<ol class=\\"a\\"><li>foo</li></ol><ol class=\\"b\\"><li>bar</li></ol>"
[FAIL] [["delete",""]] "<ol><ol><li>foo</ol><li>{}<br></li><ol><li>bar</ol></ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li><li>bar</li></ol></ol>" but got "<ol><ol><li>foo</li></ol><ol><li>bar</li></ol></ol>"
[FAIL] [["delete",""]] "<ol><li>foo[</ol>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>baz</li></ol>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<ol><li>foo[</ol><p>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>baz</li></ol>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<ol><li>foo[</ol><p>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>baz</li></ol>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<ol><li><p>foo[</ol><p>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p></li><li>baz</li></ol>" but got "<ol><li><p>foo</p></li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<ol><li><p>foo[</ol><p>bar]<ol><li>baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><p>foo</p></li><li>baz</li></ol>" but got "<ol><li><p>foo</p></li></ol><ol><li>baz</li></ol>"
[FAIL] [["delete",""]] "<ol><li>foo</ol><p>[bar<ol><li>]baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>baz</p>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<ol><li>foo</ol><p>[bar<ol><li><p>]baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>baz</p>" but got "<ol><li>foo</li></ol><ol><li><p>baz</p></li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<ol><li>foo</ol><p>[bar<ol><li><p>]baz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li></ol><p>baz</p>" but got "<ol><li>foo</li></ol><ol><li><p>baz</p></li></ol>"
[FAIL] [["delete",""]] "<ol><ol><li>foo[</ol><li>bar</ol>baz]<ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><ol><li>foo</li></ol><li>quz</li></ol>" but got "<ol><ol><li>foo</li></ol></ol><ol><li>quz</li></ol>"
[FAIL] [["delete",""]] "<ul><li>foo</ul>{}<br><ul><li>bar</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li></ul>" but got "<ul><li>foo</li></ul><ul><li>bar</li></ul>"
[FAIL] [["defaultparagraphseparator","div"],["delete",""]] "<ul><li>foo</ul><p>{}<br></p><ul><li>bar</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li></ul>" but got "<ul><li>foo</li></ul><ul><li>bar</li></ul>"
[FAIL] [["defaultparagraphseparator","p"],["delete",""]] "<ul><li>foo</ul><p>{}<br></p><ul><li>bar</ul>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo</li><li>bar</li></ul>" but got "<ul><li>foo</li></ul><ul><li>bar</li></ul>"
[FAIL] [["delete",""]] "<ol><li>foo[<li>bar]</ol><ol><li>baz</ol><ol><li>quz</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo</li><li>baz</li><li>quz</li></ol>" but got "<ol><li>foo</li></ol><ol><li>baz</li></ol><ol><li>quz</li></ol>"
[FAIL] [["delete",""]] "<div><b>[foo]</b></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div>" but got "<br>"
[FAIL] [["delete",""]] "foo<img contenteditable=false src=/img/lion.svg>[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "foo<img contenteditable=\\"false\\" src=\\"/img/lion.svg\\">bar"
Harness: the test ran to completion.
